if isDuplicate handling
isDuplicate was added to prevent one mac address from being added to the
deviceList array in multiple places.  This was a great idea to prevent
wasting memory and printing unneeded lines, however, it can be futher
improved by returning the index for that device and allowing the entry
to be updated instead of skipped.  Without this change, devices are only
logged once no matter how many times they are seen; this loses valuable
information such as "the signal was stronger at point X than at point
Y".

Additionally this commit removes a few delays which I found were simply
delaying start and didn't seem to have a purpose beyond that; everything
is still readable without.

I increased the page turn delay, because 2 seconds just isn't long
enough to see 31 devices, so I made it 3 seconds; read fast :-)

I left comments in with my lines for troubleshooting memory usage.  It
didn't turn out to be the issue I was tracking, but I figured it was a
useful comment to leave.