fix issue #6 in the cheapest way
based on a comment from @lukeswitz
https://github.com/colonelpanichacks/m5paper-Wardriver/pull/7#issuecomment-2364762344

The easiest way to not crash it to simply restart the index numbers and
write over the oldest devices in a first-in-first-out fashion.  This
isn't perfect, as it's writing over the first seen stuff without caring
if it's still seen.  This won't cause duplicates as isDuplicate will
protect us, but it may cause in range devices to shift index in the
deviceList array.  I don't think that will hurt anything though.

I also shrank the array size from 750 to 500.  I'm not sure where the
balance point is.  With 500 devices it takes 51 seconds to show the scan
results because it's 3 seconds per page and 17 pages.  That's 51 seconds
we are not scanning.  That said, a scan result *could* include more than
500 devices and then we won't display all the results.  I think the best
solution here is to swap the wifi scanning to non-blocking like the
bluetooth so we can have the best of both.