add a mutex for deviceList access
when in range of lots of ble devices (crowds of people) the code would
crash due to the display reading from deviceList while the BLE Callback
was writing deviceList.  With this mutex in place I am stable tracking
825 devices and displaying the last 2m worth of scan data.

Also a minor refactor of the battery code.

Added some debug prints and ifdef'd them out so they are easy to use but
don't affect non-debug use.